📡 Upload to npm as module
