+++ .gitignore
@@ -1,2 +1,4 @@
 node_modules
 public
+.DS_Store
+package-lock.json
