+++ .npmignore
@@ -1,0 +1,16 @@
+.*.swp
+._*
+.DS_Store
+.git
+.hg
+.npmrc
+.lock-wscript
+.svn
+.wafpickle-*
+config.gypi
+CVS
+npm-debug.log
+public
+package-lock.json
+example
+.gitignore
