+++ example/src/main.js
