+++ package.json
@@ -1,22 +1,23 @@
 {
-  "name": "svelte-multistep-forms",
-  "version": "0.0.1",
+  "name": "svelte-multistep-form",
+  "main": "public/build/bundle.js",
+  "module": "public/build/bundle.mjs",
+  "svelte": "src/index.js",
+  "version": "0.0.3",
   "scripts": {
     "build": "rollup -c",
-    "dev": "rollup -c -w",
-    "run-examples": "sirv public/examples",
-    "start": "sirv public"
+    "dev": "rollup -c -w"
   },
   "devDependencies": {
     "@rollup/plugin-commonjs": "11.0.2",
     "@rollup/plugin-node-resolve": "^7.0.0",
-    "rollup": "^1.20.0",
-    "rollup-plugin-livereload": "^1.0.0",
-    "rollup-plugin-svelte": "^5.0.3",
-    "rollup-plugin-terser": "^5.1.2",
-    "svelte": "^3.0.0"
+    "rollup": "^1.32.1",
+    "rollup-plugin-node-resolve": "^5.2.0",
+    "rollup-plugin-svelte": "^5.2.2",
+    "svelte": "^3.23.0"
   },
-  "dependencies": {
-    "sirv-cli": "^0.4.4"
+  "homepage": "",
+  "bugs": {
+    "url": "https://github.com/edenriquez/svelte-multistep-form/issues"
   }
 }
